+++ Assets/Package/Runtime/Scripts/Modules/ContainerModuleLoader.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using Autofac;
@@ -27,6 +26,10 @@ namespace MrWatts.Internal.FuelInject
         private bool automaticallyAddRootGameObjectModules = true;
 
         [SerializeField]
+        [Tooltip("Whether to automatically load the Unity logger (sub)module. You can disable this if you don't want to load UnityLoggerModule, or want to load it yourself to perform additional customizations to it.\n\nFuel Inject needs an IUnityKernelLogger to function correctly, so you are responsible for binding one yourself if you choose not to use the built-in UnityLoggerModule!")]
+        private bool automaticallyLoadUnityLoggerModule = true;
+
+        [SerializeField]
         [Tooltip("Whether to enable container diagnostics. This will automatically load the ContainerDiagnosticsModule for the current container, logging its output to the Unity console. If you want to log somewhere else, disable this and load the module yourself with a custom tracer.")]
         private bool enableContainerDiagnostics = false;
 
@@ -41,28 +44,6 @@ namespace MrWatts.Internal.FuelInject
             LoadModules(builder);
 
             builder.RegisterSource(new OrderedEnumerableRegistrationSource());
-            builder.RegisterComposite<CompositeInitializable, IInitializable>().SingleInstance();
-            builder.RegisterComposite<CompositeAsyncInitializable, IAsyncInitializable>().SingleInstance();
-            builder.RegisterComposite<CompositeTickable, ITickable>().SingleInstance();
-            builder.RegisterComposite<CompositeFixedTickable, IFixedTickable>().SingleInstance();
-            builder.RegisterComposite<CompositeLateTickable, ILateTickable>().SingleInstance();
-            builder.RegisterComposite<CompositeDisposable, IDisposable>().SingleInstance();
-            // builder.RegisterComposite<CompositeAsyncDisposable, IAsyncDisposable>();
-
-            builder.RegisterType<ObjectDependencyInjector>()
-                .AsSelf()
-                .As<IInjector<object>>()
-                .SingleInstance();
-
-            builder.RegisterType<UnityGameObjectDependencyInjector>()
-                .AsSelf()
-                .As<IInjector<GameObject>>()
-                .SingleInstance();
-
-            builder.RegisterType<UnitySceneDependencyInjector>()
-                .AsSelf()
-                .As<IInjector<Scene>>()
-                .SingleInstance();
 
             builder.RegisterType<ConfigurableComponentContextProxy>()
                 .AsSelf()
@@ -94,6 +75,14 @@ namespace MrWatts.Internal.FuelInject
             }
 
             List<(int Priority, IModule Module)> modules = RetrieveStartupModules();
+
+            builder.RegisterModule(new KernelModule());
+            builder.RegisterModule(new UnityInjectionModule());
+
+            if (automaticallyLoadUnityLoggerModule)
+            {
+                builder.RegisterModule(new UnityLoggerModule());
+            }
 
             if (automaticallyAddRootGameObjectModules)
             {